requirements.txt
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
tqdm>=4.66.0
scikit-learn>=1.3.0
ultralytics>=8.2.0
timm>=0.9.7
pyyaml>=6.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
opencv-python>=4.8.0

